requirements.txt
pdfplumber
pandas
PyPDF2
openpyxl
jpype1
sqlalchemy
psycopg2-binary
pg8000
python-dotenv
tabula
